rename payments and payouts, fix payout value if string, add documentation
