+++ examples/06_websocket.html
@@ -2,7 +2,7 @@
 <html>
 
 <head>
-    <title>Socket.IO chat</title>
+    <title>Socket.IO</title>
     <style>
         li {
             font-size: 2em;
