+++ index.js
@@ -1,5 +1,5 @@
 module.exports = require('./lib/create-server')
-module.exports.payout = require('./lib/payouts')
-module.exports.payments = require('./lib/payments')
+module.exports.payout = require('./lib/payout')
+module.exports.payment = require('./lib/payment')
 let { getBalance } = require('./lib/Account')
 module.exports.getBalance = getBalance
