+++ lib/Account.js
@@ -57,6 +57,21 @@ function getNewAddress() {
   })
 }
 
+/**
+ * Returns the balance
+ * @returns {number} balance
+ * @example
+ * // get payout by txhash
+ * paymentModule.getBalance()
+ *  .then(balance => {
+ *      console.log(balance)
+ *    })
+ *  .catch(err => {
+ *    console.log(err)
+ *  })
+ * //example balance:
+ * 15
+ */
 function getBalance() {
   return new Promise(async (resolve, reject) => {
     let payments = db.get('payouts').value()
