+++ lib/WebSockets.js
@@ -58,3 +58,13 @@ module.exports = {
   emit,
   start
 }
+
+/**
+ * WebSockets with socket.io
+ *
+ * backend:
+ * https://github.com/machineeconomy/iota-payment/blob/master/examples/06_websockets.js
+ *
+ * frontend:
+ * https://github.com/machineeconomy/iota-payment/blob/master/examples/06_websocket.html
+ */
