+++ lib/paymentHandler.js
@@ -2,7 +2,7 @@ module.exports = { start }
 
 const { getCurrentPayoutIndex } = require('./Account')
 const iotaCore = require('@iota/core')
-const { getOpenPayments } = require('./payments')
+const { getOpenPayments } = require('./payment')
 const db = require('./Database')
 const eventHandler = require('./eventHandler')
