+++ lib/payout.js
@@ -167,7 +167,45 @@ function generateAddresses(start_index, end_index) {
   return addresses
 }
 
-//create payout
+/**
+ * Creates and returns a payout.
+ * @param {object} payoutData - data for the payout
+ * @param {address} payoutData.address - 90 trytes iota address (with checksum)
+ * @param {number} payoutData.amount - amount of iotas
+ * @param {string} [payoutData.message] - message which will be send with the transaction
+ * @param {string} [payoutData.tag] - tryte tag
+ * @param {number} [payoutData.startIndex] - custom start index to search for inputaddresses
+ * @param {number} [payoutData.endIndex] - custom end index to search for inputaddresses
+ * @returns {object} payout
+ * @example
+ * // create a payout
+ * let payoutObject = {
+ *   //required
+ *   address: 'HW99PKRDWBUCCWLEQDONQXW9AXQHZAABYEKWSEFYAUIQWWCLVHAUJEQAZJACAGPZBZSQJIOUXRYYEXWZCXXOAJMZVY',
+ *   value: 1,
+ *   //optional
+ *   message: 'Example message',
+ *   tag: 'TRYTETAG',
+ *   //indexes for input addresses, only required in special cases
+ *   // startIndex: 0,
+ *   // endIndex: 1
+ * }
+ * paymentModule.payout.send(payoutObject)
+ *  .then(payout => {
+ *    console.log(payout)
+ *  })
+ *  .catch(err => {
+ *    console.log(err)
+ *  })
+ * //example payout:
+ * { address:
+ *  'HW99PKRDWBUCCWLEQDONQXW9AXQHZAABYEKWSEFYAUIQWWCLVHAUJEQAZJACAGPZBZSQJIOUXRYYEXWZCXXOAJMZVY',
+ * value: 1,
+ * message: 'Example message',
+ * tag: 'TRYTETAG',
+ * payed: false,
+ * id: '1570619992125' }
+ */
 function send(payoutData) {
   return new Promise(async function(resolve, reject) {
     try {
@@ -200,6 +238,37 @@ function send(payoutData) {
   })
 }
 
+/**
+ * Returns all payouts
+ * @returns {Object[]} payouts
+ * @example
+ * // get payouts
+ * paymentModule.payout.getPayouts()
+ *  .then(payouts => {
+ *    console.log(payouts)
+ *  })
+ *  .catch(err => {
+ *    console.log(err)
+ *  })
+ * //example payouts:
+ * [
+ *  {
+ *    "address": "USKSGFKPWVXD9EUOJXWNLVWG9FDTIPGBUYB9BQNMMUWIEOMFNDCCUKCJMEPRICBHZNRAIZFGNPK9GCGBYQAEWNJRMC",
+ *    "value": 1,
+ *    "message": "Example message",
+ *    "payed": true,
+ *    "id": "1570616325073",
+ *    "txhash": "XKLLL9B9AUN9EASCAQAQHEYDLLEUDDUCOTJVNTJUSZVTNWUTDPHZUFUJAHFZJOSQYYEPJSWRDXDJ99999"
+ *  },
+ *  {
+ *    "address": "HW99PKRDWBUCCWLEQDONQXW9AXQHZAABYEKWSEFYAUIQWWCLVHAUJEQAZJACAGPZBZSQJIOUXRYYEXWZCXXOAJMZVY",
+ *    "value": 1,
+ *    "message": "Example message",
+ *    "payed": false,
+ *    "id": "1570616561382"
+ *  }
+ * ]
+ */
 function getPayouts() {
   return new Promise(function(resolve, reject) {
     resolve(db.get('payouts').value())
@@ -217,6 +286,29 @@ function updatePayout(payout) {
   })
 }
 
+/**
+ * Returns open payouts
+ * @returns {Object[]} open payouts
+ * @example
+ * // get open payouts
+ * paymentModule.payout.getOpenpayouts()
+ *  .then(payouts => {
+ *    console.log(payouts)
+ *  })
+ *  .catch(err => {
+ *    console.log(err)
+ *  })
+ * //example payouts:
+ * [
+ *  {
+ *    "address": "HW99PKRDWBUCCWLEQDONQXW9AXQHZAABYEKWSEFYAUIQWWCLVHAUJEQAZJACAGPZBZSQJIOUXRYYEXWZCXXOAJMZVY",
+ *    "value": 1,
+ *    "message": "Example message",
+ *    "payed": false,
+ *    "id": "1570616561382"
+ *  }
+ * ]
+ */
 function getOpenpayouts() {
   return new Promise(function(resolve, reject) {
     resolve(
@@ -228,6 +320,26 @@ function getOpenpayouts() {
   })
 }
 
+/**
+ * Returns payout by id
+ * @returns {Object} payment
+ * @example
+ * // get payout by id
+ * paymentModule.payout.getPayoutByID('1570611186704')
+ *  .then(payout => {
+ *    console.log(payout)
+ *  })
+ *  .catch(err => {
+ *    console.log(err)
+ *  })
+ * //example payout:
+ * { address:
+ *  'USKSGFKPWVXD9EUOJXWNLVWG9FDTIPGBUYB9BQNMMUWIEOMFNDCCUKCJMEPRICBHZNRAIZFGNPK9GCGBYQAEWNJRMC',
+ * value: 1,
+ * message: 'Example message',
+ * payed: false,
+ * id: '1570611186704' }
+ */
 function getPayoutByID(id) {
   return new Promise(function(resolve, reject) {
     const Payout = db
@@ -238,6 +350,28 @@ function getPayoutByID(id) {
   })
 }
 
+/**
+ * Returns payout by txhash
+ * @returns {Object} payment
+ * @example
+ * // get payout by txhash
+ * paymentModule.payout.getPayoutByTxhash('XKLLL9B9AUN9EASCAQAQHEYDLLEUDDUCOTJVNTJUSZVTNWUTDPHZUFUJAHFZJOSQYYEPJSWRDXDJ99999')
+ *  .then(payout => {
+ *   console.log(payout)
+ *  })
+ *  .catch(err => {
+ *     console.log(err)
+ *   })
+ * //example payout:
+ * { address:
+ *  'USKSGFKPWVXD9EUOJXWNLVWG9FDTIPGBUYB9BQNMMUWIEOMFNDCCUKCJMEPRICBHZNRAIZFGNPK9GCGBYQAEWNJRMC',
+ * value: 1,
+ * message: 'Example message',
+ * payed: true,
+ * id: '1570616325073',
+ * txhash:
+ *  'XKLLL9B9AUN9EASCAQAQHEYDLLEUDDUCOTJVNTJUSZVTNWUTDPHZUFUJAHFZJOSQYYEPJSWRDXDJ99999' }
+ */
 function getPayoutByTxhash(txhash) {
   return new Promise(function(resolve, reject) {
     const Payout = db
@@ -277,7 +411,7 @@ function payoutValidator({
       ) {
         throw 'Invalid iota value'
       }
-      payout.value = value
+      payout.value = parseInt(value)
 
       if (typeof startIndex != 'undefined') {
         payout.startIndex = startIndex
