+++ lib/payoutHandler.js
@@ -5,7 +5,7 @@ const {
   sendPayout,
   updatePayout,
   payoutValidator
-} = require('./payouts')
+} = require('./payout')
 const db = require('./Database')
 const eventHandler = require('./eventHandler')
