+++ lib/routes.js
@@ -1,7 +1,7 @@
 module.exports = createRoutes
 
-const { createPayment, getPayments, getPaymentByID } = require('./payments.js')
-const { getPayouts, getPayoutByID } = require('./payouts.js')
+const { createPayment, getPayments, getPaymentByID } = require('./payment.js')
+const { getPayouts, getPayoutByID } = require('./payout.js')
 const { getBalance } = require('./Account.js')
 
 function createRoutes(app, mount = 'payments', value = 0) {
